Relocating the hack to suspend client tasks when the journal is overextended to the DataServer (rather than the DataService).
Added logging when blocked/released per task.

git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/DEV_BRANCH_27_OCT_2009@2291 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/journal/ConcurrencyManager.java
@@ -1204,7 +1204,7 @@ public class ConcurrencyManager implements IConcurrencyManager {
              * 
              * @todo convert to WARN
              */
-            log.error("overextended=" + overextension + "x : "
+            log.error("overextended=" + (int)overextension + "x : "
                     + task.toString());
 
         }
